pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]

[project]
name = "matrix-python"
version = "0.1.2-alpha"
authors = [
    { name="Simon Roy" },
    { name=" Chris Dedman Rollet " }
]
maintainers = [
    { name="Code Society Lab", email="admin@codesociety.xyz" }
]
description = "An easy-to-use Matrix bot framework designed for quick development and minimal setup"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"

dependencies = [
    "matrix-nio",
    "logger",
    'PyYAML',
    'markdown',
    'APScheduler',
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "mypy",
    "types-PyYAML",
    'types-Markdown',
]

[project.urls]
"Homepage" = "https://codesociety.xyz"
"Source" = "https://github.com/Code-Society-Lab/matrixpy"
"Issues" = "https://github.com/Code-Society-Lab/matrixpy/issues"

[tool.setuptools]
packages = ["matrix"]
license-files = []